Fixed merge conflicts bug and some refactoring to improve code.
