+++ src/main/resources/view/PersonOverview.fxml
@@ -53,6 +53,7 @@
                                 <Label text="City" GridPane.rowIndex="3" />
                                 <Label text="Postal Code" GridPane.rowIndex="4" />
                                 <Label text="Birthday" GridPane.rowIndex="5" />
+                                <Label text="Contact Group" GridPane.rowIndex="6" />
                                 <Label fx:id="firstNameLabel" styleClass="label-bright" text="Label" GridPane.columnIndex="1" />
                                 <Label fx:id="lastNameLabel" styleClass="label-bright" text="Label" GridPane.columnIndex="1" GridPane.rowIndex="1" />
                                 <Label fx:id="streetLabel" styleClass="label-bright" text="Label" GridPane.columnIndex="1" GridPane.rowIndex="2" />
